modularize more, separate function desugar and romgen
